Use event_pending to check if an event is pending or registered under MSVC
I've only made this change for MSVC because I don't have the libevent server working properly under MSVC yet.
Under MSVC at least, `isEventRegistered` was returning `true` when it shouldn't have been, prompting me to ask in the IRC channel for libevent, and, after asking a few times over a few days, I finally got a response that said that `event_pending` is the correct way to be doing this, so that's what MSVC will now use.